# RegoLinkFlowConnector/Rego.LinkConnector.API/Assets/RegoLinkCustomConnector_OPENAPI.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "RegoLink for CA PPM",
    "description": "RegoLink Flow Connector"
  },
  "host": "cappmflowdev.azurewebsites.net",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/RegoLinkFlowConnector/GetActions": {
      "get": {
        "description": "List all available actions",
        "summary": "GetActions",
        "operationId": "GetActions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetActions"
            }
          },
          "500": {
            "description": "ERROR",
            "schema": {
              "$ref": "#/definitions/ResponseError"
            }
          }
        },
        "x-ms-visibility": "internal"
      }
    },
    "/api/RegoLinkFlowConnector/GetParameters/{actionID}": {
      "get": {
        "description": "List all parameters based on the selected Action",
        "summary": "GetParameters",
        "operationId": "GetParametersSchema",
        "parameters": [
          {
            "name": "actionID",
            "type": "string",
            "in": "path",
            "description": "Select Actions",
            "required": true,
            "x-ms-summary": "Actions"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object"
            }
          },
          "500": {
            "description": "ERROR",
            "schema": {
              "$ref": "#/definitions/ResponseError"
            }
          }
        },
        "x-ms-visibility": "internal"
      }
    },
    "/api/RegoLinkFlowConnector/ExecuteAction/{actionID}": {
      "post": {
        "description": "Execute Action",
        "summary": "Execute Action",
        "operationId": "ExecuteAction",
        "parameters": [
          {
            "name": "actionID",
            "type": "string",
            "in": "path",
            "description": "Select Actions",
            "required": true,
            "x-ms-summary": "Actions",
            "x-ms-dynamic-values": {
              "operationId": "GetActions",
              "value-path": "Id",
              "value-title": "Name"
            }
          },
          {
            "name": "dynamicActionSchema",
            "in": "body",
            "description": "Dynamic Schema of items in selected action",
            "schema": {
              "type": "object",
              "x-ms-dynamic-schema": {
                "operationId": "GetParametersSchema",
                "parameters": {
                  "actionID": {
                    "parameter": "actionID"
                  }
                },
                "value-path": "items"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "ERROR",
            "schema": {
              "$ref": "#/definitions/ResponseError"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "GetActions": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string"
          },
          "Name": {
            "type": "string"
          }
        },
        "required": [
          "Id",
          "Name"
        ]
      }
    },
    "ResponseError": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      },
      "required": [
        "code",
        "message"
      ]
    },
    "DynamicResponseGetParametersSchema": {
      "type": "object",
      "x-ms-dynamic-schema": {
        "operationId": "GetParametersSchema",
        "parameters": {
          "actionID": {
            "parameter": "actionID"
          }
        },
        "value-path": "items"
      }
    }
  },
  "parameters": {},
  "responses": {},
  "securityDefinitions": {
    "basic_auth": {
      "type": "basic"
    }
  },
  "security": [],
  "tags": []
}